+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # ActFramework Change Log
 
 **1.9.2**
+* Add "The Wall of Coding Wisdoms" into default Zen list #1388
 * Response content type get overridden - case 2 #1387
 * Response content type get overridden #1386
 * Add act version info into API doc page and test report page #1380